% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ques_data.R
\name{match_ques_data}
\alias{match_ques_data}
\title{Binds a qeustionnaire with a data.df}
\usage{
match_ques_data(data.df, questionnaire.name)
}
\arguments{
\item{questionnaire.name}{}
}
\description{
Binds a qeustionnaire with a data.df
}
